pastel/requirements.txt
pre-commit
yapf
pylint

tensorboardX
numpy>=1.20
Pillow
matplotlib
yacs
wandb
seaborn
tqdm
timm==0.9.7
opencv-python==4.5.4.60
pytest-shutil
mmcv==1.6.2
scikit-image
faiss-gpu
kornia
pytorch-lightning[extra]==1.8.6
lightning-utilities==0.5.0
omegaconf==2.3.0
scikit-learn==1.2.2
numba==0.54.1
cupy-cuda111